Fixed #40
When setting a custom entity's attributes, the DefaultAttributeContainer$Builder object should be registered with the entity type to the DefaultAttributeRegistry via Fabric API, and as such handled like vanilla.

Since LivingEntity already handles access to the AttributeContainer, the getAttributes() method should not be overridden, as properly registering a DefaultAttributeContainer to the entity type as described above is sufficient.

Regardless of the incompatibility with Data Attributes, it is not optimal to assign an AttributeContainer to an entity on the first time it is needed via a  null check - it is better to do this when the entity is initialised/created.